third_party/WebKit/LayoutTests/external/wpt/resource-timing/resource_timing_store_and_clear_during_callback-expected.txt

This is a testharness.js-based test.
PASS No entry should be stored in resource timing buffer since its cleared once an item arrived.
FAIL 6 resource timing entries should be moved to global buffer. assert_equals: 6 resource timing entries should be moved to global buffer. expected 6 but got 7
FAIL http://web-platform.test:8001/resource-timing/resources/empty.js is not expected to be in the Resource Timing buffer assert_unreached: Reached unreachable code
PASS http://web-platform.test:8001/resource-timing/resources/empty_script.js is expected to have initiatorType script
PASS http://web-platform.test:8001/resource-timing/resources/resource_timing_test0.js is expected to have initiatorType script
PASS http://web-platform.test:8001/resource-timing/resources/webperftestharness.js is expected to have initiatorType script
PASS http://web-platform.test:8001/resource-timing/resources/webperftestharnessextension.js is expected to have initiatorType script
PASS http://web-platform.test:8001/resources/testharness.js is expected to have initiatorType script
PASS http://web-platform.test:8001/resources/testharnessreport.js is expected to have initiatorType script
Harness: the test ran to completion.
